added action menu to jobs details menu and fix some bugs
