--- src/common/TextField/textField.scss
@@ -1,17 +1,0 @@
-@import '../../scss/colors.scss';
-@import '../../scss/borders.scss';
-
-.text_field_container {
-  display: flex;
-  max-width: 200px;
-  align-items: center;
-  border: $tableFilterBorder;
-  border-radius: 4px;
-  margin-right: 15px;
-  margin-bottom: 15px;
-  .input {
-    height: 38px;
-    width: 200px;
-    padding: 10px 10px 10px 64px;
-  }
-}
